pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Basic MCP server with SSE connection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.0"
]
